# Speed ups to ess_rhat