This PR adds some speed-ups to `mean_autocov`, which in turn speeds up `ess_rhat` for two ess methods.

### `ESSMethod`:

`eltype(samples) isa LinearAlgebra.BlasReal` is a malformed check and will always be false. We could change it to `eltype(samples) <: LinearAlgebra.BlasReal`, but the code within it calls a `dot` method that doesn't exist. Besides, for `StridedArray` inputs with `BlasReal` eltype, the second call to `dot` hits https://github.com/JuliaLang/julia/blob/v1.6.1/stdlib/LinearAlgebra/src/matmul.jl#L9, which forwards to `BLAS.dot` already, so this check is not necessary.

And it turns out removing the check gives us an almost 2x speed-up and cuts allocations substantially:

```julia
julia> using InferenceDiagnostics

julia> x = randn(1000, 10, 4);

julia> @btime ess_rhat($x);
```
current:
```julia
  222.542 μs (2759 allocations: 119.52 KiB)
```

this pr:
```julia
  140.919 μs (7 allocations: 31.95 KiB)
```

### `BDAESSMethod`:

Replacing a `sum` over a difference with 3 equivalent calls to `dot` enables us to get almost a 2x speed-up. The speed-up is probably due to hitting `BLAS.dot`.

```julia
julia> @btime ess_rhat($x; method=BDAESSMethod())
```

current:
```julia
  429.368 μs (56 allocations: 32.73 KiB)
```

this pr:
```julia
  220.347 μs (8 allocations: 31.98 KiB)
```