@sethaxen 2021-06-22
## Updated benchmarks

#### Setup

```julia
using InferenceDiagnostics, BenchmarkTools, Random
rng = MersenneTwister(42)
samples = randn(rng, 1000, 8)
chain_var = rand(rng, 8)
esscache = InferenceDiagnostics.ESSCache(samples, chain_var)
bdaesscache = InferenceDiagnostics.BDAESSCache(samples, chain_var, 0.0)
```

### `ESSMethod`

```julia
julia> @btime $(InferenceDiagnostics.mean_autocov)(1, $esscache);
```

#### Julia 1.0

current:
```julia
  2.529 μs (42 allocations: 1.52 KiB)
```

this pr:
```julia
  1.686 μs (17 allocations: 1.05 KiB)
```

#### Julia 1.6

current:
```julia
  3.256 μs (53 allocations: 1.78 KiB)
```

this pr:
```julia
  1.765 μs (0 allocations: 0 bytes)
```

### `ESSMethod`

```julia
julia> @btime $(InferenceDiagnostics.mean_autocov)(1, $bdaesscache);
```

#### Julia 1.6

current:
```julia
  12.282 μs (1 allocation: 16 bytes)
```

this pr:
```julia
  1.992 μs (1 allocation: 16 bytes)
```