@sethaxen 2021-06-23
Hmm, isn't the format workflow supposed to review with suggested changes instead of just failing? I'll format locally. Btw, @devmotion would you mind adding me as a collaborator on this repo?