@devmotion 2021-06-23
It opens suggestions only if the PR is not from a fork (due to Github's security restrictions for actions), and prints the suggestions in the logs of the action otherwise.