@devmotion 2021-06-29
It seems recent changes in MLJ (refactoring of (MLJ)ScientificTypes) broke the MLJ integration.